requirements.txt
pydantic
ujson
tqdm
colorlog
dumb-menu
tenacity
requests
aiohttp
pycryptodome
UnityPy
